+++ lib/app_modules/med_user/med_dashboard.dart
@@ -3,9 +3,10 @@ import 'package:flutter_svg/svg.dart';
 import 'package:get/get.dart';
 import 'package:mboacare/app_modules/med_user/screen/dashboard/blog/blog.dart';
 import 'package:mboacare/app_modules/med_user/screen/dashboard/home.dart';
-import 'package:mboacare/app_modules/med_user/screen/dashboard/hospital/hospital_page.dart';
+
 import 'package:mboacare/app_modules/med_user/screen/dashboard/settings.dart';
 import 'package:mboacare/app_modules/notifications/notifications.dart';
+import 'package:mboacare/app_modules/user/screens/dashboard/hospital.dart';
 import 'package:mboacare/global/styles/assets_string.dart';
 import 'package:mboacare/global/styles/colors.dart';
 import 'package:mboacare/widgets/drawer_widget.dart';
@@ -23,7 +24,7 @@ class _MedDashboardState extends State<MedDashboard> {
 
   final List<Widget> _pages = [
     const Home(),
-    const HospitalPage(),
+    const HospitalDashboard(),
     const Blog(),
     const Settings(),
   ];
