+++ lib/app_modules/med_user/screen/dashboard/home.dart
@@ -1,9 +1,7 @@
-
-
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mboacare/app_modules/med_user/screen/dashboard/blog/blog.dart';
-import 'package:mboacare/app_modules/med_user/screen/dashboard/hospital/hospital_page.dart';
+import 'package:mboacare/app_modules/user/screens/dashboard/hospital.dart';
 import 'package:mboacare/global/styles/appStyles.dart';
 import 'package:mboacare/global/styles/assets_string.dart';
 import 'package:mboacare/global/styles/colors.dart';
@@ -89,7 +87,7 @@ class _HomeState extends State<Home> {
                       subtitle: 'Browse through facilities',
                       iconImage: ImageAssets.hospital,
                       onTap: () {
-                        Get.to(() => const HospitalPage());
+                        Get.to(() => const HospitalDashboard());
                       },
                     ),
                     HomeNavigationItems(
