+++ lib/app_modules/med_user/screen/dashboard/settings.dart
@@ -125,7 +125,7 @@ class Settings extends StatelessWidget {
                   subtitle: 'Manage Facilities',
                   trailingIcon: const Icon(Icons.arrow_forward_ios_outlined),
                   onTap: () {
-                    Navigator.pushNamed(context, '/facilities');
+                    Navigator.pushNamed(context, '/manageFacilities');
                   }),
               SettingsPageListTiles(
                   icon: const Icon(
