+++ lib/main.dart
@@ -1,10 +1,15 @@
+import 'dart:io';
+
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:mboacare/app_modules/auth/forgotPassword.dart';
 import 'package:mboacare/app_modules/auth/checkMail.dart';
 import 'package:mboacare/app_modules/med_user/screen/dashboard/home.dart';
-import 'package:mboacare/app_modules/med_user/screen/dashboard/hospital/add_facilities_page.dart';
+
+import 'package:mboacare/app_modules/med_user/screen/dashboard/hospital/add_hopital.dart';
 import 'package:mboacare/app_modules/med_user/screen/dashboard/hospital/awaitApprovalScreen.dart';
+import 'package:mboacare/app_modules/med_user/screen/dashboard/hospital/edit_facilities_page.dart';
+import 'package:mboacare/app_modules/med_user/screen/dashboard/hospital/manageFacilities.dart';
 import 'package:mboacare/app_modules/med_user/screen/inner_screen/user_profile_page.dart';
 import 'package:mboacare/app_modules/user/screens/inner_screen/aboutUs.dart';
 import 'package:mboacare/global/theme/themeConstants.dart';
@@ -104,7 +109,7 @@ class MyApp extends StatelessWidget {
         }
         return supportedLocales.first;
       },
-      home: const SplashScreen(),
+      home: const ManageFacilities(),
       routes: {
         ('/home'): (context) => Home(),
         ('/aboutUs'): (context) => AboutUs(),
@@ -113,7 +118,15 @@ class MyApp extends StatelessWidget {
         ('/awaitApproval'): (context) => AwaitApprovalScreen(),
         ('/resetPassword'): (context) => CheckMailScreen(),
         ('/forgotPassword'): (context) => ForgotPasswordScreen(),
-        ('/addFacilities'): (context) => AddFacilitiesPage(),
+
+        ('/editFacilities'): (context) => EditFacilitiesPage(),
+        ('/manageFacilities'): (context) => ManageFacilities(),
+        ('/addHospital'): (context) => AddHospitalPage(
+              placeName: Platform.localeName,
+            ),
+        //('/addHospital'): (context) => AddHospitalPage(
+        //  placeName: Platform.localeName,
+        //  )
       },
     );
   }
